requirements.txt
selenium==4.17.2
sqlalchemy==2.0.26
psycopg2-binary==2.9.9
python-dotenv==1.0.1